Move detailed Figure Caption container snapshot tests from Figure to Caption container
